Added camera shake effect and reusable fading objects
I added camera shakes that can be controlled through the camera manager with a duration and strength. Also added a script called special objects that is able to fade in and out using a animation curve. If it automatically disappears when you enter the collider its probably because you have the curve in the opposite direction.